API simplification
