+++ src/index.js
@@ -1,10 +1,4 @@
 // Classic web-mercator-project
 export {default as default} from './flat-viewport';
-export {default as ViewportMercatorProject} from './flat-viewport';
 export {default as FlatViewport} from './flat-viewport';
-
-// 3D viewports
-export {default as Viewport} from './viewport';
 export {default as PerspectiveViewport} from './perspective-viewport';
-export {default as OrthographicViewport} from './orthographic-viewport';
-export {default as WebMercatorViewport} from './web-mercator-viewport';
